pyproject.toml
[project]
name = "schematic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "boto3>=1.38.23",
    "fastapi[standard]>=0.115.12",
    "genson>=1.3.0",
    "nltk>=3.9.1",
    "pyspellchecker>=0.8.3",
    "python-decouple>=3.8",
    "python-levenshtein>=0.27.1",
    "sqlalchemy>=2.0.41",
    "xmltodict>=0.14.2",
]
